+++ CHANGES.md
@@ -4,6 +4,10 @@ This is a record of all past `gettsim` releases and what went into them in rever
 chronological order. We follow [semantic versioning](https://semver.org/) and all
 releases are available on [Anaconda.org](https://anaconda.org/conda-forge/gettsim).
 
+## unreleased
+
+- {gh}`1064` Improve paths handling, small fixes ({ghuser}`hmgaudecker`).
+
 ## v1.0a — 2025-07-27
 
 - {gh}`1062` Fix some end dates ({ghuser}`hmgaudecker`)